add `parametrize_cases` helper
factored out of `test_dataframe_io_roundtrips.py`
